Support projects only for subdirs of the project root.
For large codebases, path exclusion is unworkable (too many paths to
exclude). It's better to only include the directories you want to
include.

This change accomplishes that, by changing the .projectile file so
that entries prefixed with + are subdirectory includes, - are
subdirectory and file pattern excludes, and no prefix are also
subdirectory and file pattern excludes, to preserve backwards
compatibility.