Name the default helm-projectile file action
When operating on .el files, Helm recklessly appends an action to the
list of actions, which seems to expect the extended format for actions,
otherwise hitting RET on an emacs lisp file in helm-projectile causes
Emacs to complain. This change gives the default action the name "Find
File", and fixes this problem.