allow projectile-project-root to find the root for symlinked files
This covers the case where you have a git repo ~/configs (for dotfiles
etc.), and the subdir emacs of that repo is symlinked to ~/.emacs.d

projectile-project-root in a buffer visiting ~/.emacs.d/init.el will
find the root at ~/configs.

It also covers the case of a single file symlinked outside a repo.