Fix the problem of corrupting "known project file"
At times the "known project file" (the file "projectile-bookmarks.eld") gets
damaged, i.e., the end of the list gets abbreviated with " \.\.\.". This is
caused by calling `prin1-to-string' with random current non-nil value of
`print-length'. A non-nil value of this variable determines the maximum length
of list to print before abbreviating. Therefor, by binding `print-length' to nil
temporarily while calling `prin1-to-string', we forces no-limit printing which
addresses the problem.