Allow direct modification of `projectile-command-map'
I wanted to add a second prefix key pointing at the map on C-c p. I
realized that I could duplicate the entire command map by
using (key-binding (kbd "C-c p")) as a define-key target, but if I
wanted to modify the inner map then that binding was order dependent.
Ie, I could modify the inner map by defining it twice, once
for each prefix, or modify one prefix and then add another prefix.

The inner map appears to be defined this way to be more terse, but it
interferes with traditional binding functions. Can we make it more
accessible, or is there some other reason to nest it this way?