tags: fix tags function not work on ubuntu and mac
This patch was try to fix the tags generate function and
tags find function in projectile.

in the ubuntu 12.04 and Mac OS 10, there was using etags
as the tags util.

also, the etags and ctags version change already don't
support adding exclude-file arguments. also the -Re
it will only accept a list of input file list.

So, I use projectile repo file list command and pipe
to etags to create a full tags.

Change this to make tags function work.

If there is a need to keep back compat., better add a version check.