Improve performance of projectile-find-other-file
Currently, the command searches for files in filelist
using (projectile-current-project-files) for every file. It would be
slow as the number of handlers increases. This change narrows project
files to those that only contains the filename, then perform searches on
narrowed set.

Change filter to -filter

filter should be -filter or error occurs.