Allow helm-projectile to be used outside of projects
Currently, helm-projectile contains a separate command for displaying
list of known projects, that is helm-projectile-switch-projects. But,
the command still can only work inside a project.

This change allows helm-projectile to be invoked outside of a
project. When in such scenario, helm-projectile simply displays a list
of projects for jumping. After this change, helm-projectile also does
not need to rely on projectile-switch-project and provides a better
project listing, since it can perform more than one action.