Add helm-projectile-grep/ack/ag
Add replacement commands for projectile-grep, projectile-ack and
projectile-ag. When a symbol is at point, use that symbol and search at
project root. If a region is active, use the region instead (only work
with helm-projectile-grep/ack).

helm-projectile-grep/ack ignore files/directories based on
grep-find-ignored-files and grep-find-ignored-directories variables,
along with projectile-ignored-files and projectile-ignored-directories.