git cannot run submodule if the .git folder is not at the project root folder.
projectile-vcs-root determines a project is git traversing up the path
until a `.git` folder is found.  This is fine up until the submodule
command.

If you run git submodule --quiet on a sub folder, you get

"You need to run this command from the toplevel of the working tree."

This can be problematic for projects that have a projectile-project-root
that is a child directory of its actual repository root.