[Fix #532] Treat .projectile as a generic project
We need to explicitly treat directory that contains .projectile as a
generic project. Otherwise, there will be a mismatch between
projectile-project-root and projectile-locate-dominating-file.
projectile-project-root always returns the root that contains
.projectile. But projectile-vcs currently does not check for
.projectile, and as a result when it see no .projectile file, it will
call projectile-locate-dominating-file to walk up parent directory until
it finds appropriate file to identify a project. In this case, it will
walk up the parents of projectile-project-root, and as a result if the
parent is a Git project, it will invoke the Git submodule command at the
current projectile-project-root, while the appropriate place to invoke
is the parent of projectile-project-root. This is problematic in older
Git.