Add a timer to initialize command history properly
With a timer, it leaves enough time for loading minibuffer history.

Without a timer, helm-projectile-ack/ag prevents user to quit the
commands with C-g in one try, thus requires multiple C-g pressings. A
timer introduces a brief delay for Emacs to accept user input and fix
multiple C-g pressings issue. This is also the way helm-find-files-grep
does. This is just a speculation, but it fixes the issue.