Fix helm-projectile-ack file ignoring
Currently helm-projectile-ack processes ignored files and directories
the same as projectile-ack. This works for directories, but not work for
ignoring files with wildcard matching. When user adds files with
wildcard, ack will see this not as an argument to it but a search string
and returns no result.

This is fixed by using "--ignore-file=match:" pattern that accepts a
regex that matches files. Also, shell-quote-argument must be used to
properly pass the file patterns with regex in it, otherwise ack also
returns no result.