Enable helm projectile sources reusable
Currently, the sources can only be used in a project recognized by
Projectile. It cannot be composed to be used with other commands like
helm-for-files. This change supports such use case without compromise
the consistency of Projectile: Helm Projectile commands are responsible
for throwing the same errors as stock Projectile, but not the source.