prevent infinite loop with new virtual dired name
The proposed change is not meant to be checked in.
It is merely to document the problem.  

As documented in the code comment, I can't seem to exit the minibuffer when I type a new name.
First I type C-c p f then start selecting files via hitting Control-SPC on a few files on the ensuing helm buffer.
To create a new virtual dired buffer I type C-c f which asks for a dired buffer name in the mini buffer.
I type a new name, i.e., not a pre-exising dired buffer name then hit enter.
Rather than accepting the new name, emacs goes into infinite loop with "Display not ready" error message.

The hack in this request demonstrate that infinite loop is caused in part by helm-maybe-exit-minibuffer.