Stop depending on s and f
Only two functions from `s' were used for which no equivalent function
exists in all supported Emacs versions.  So add a backward compatibility
definition of `string-suffix-p' and add `projectile-trim-string' as a
replacement for `s-trim'.

In all other cases *not* using `s' and `f' is actually an improvement,
in my humble opinion.  And having to define the two mentioned functions
does therefor not outweigh not having two additional dependencies.

Also see #623.