Fix generation of command for tag generation
Previously, command line that was being generated was this:
ctags -Re -f TAGS "--exclude=<pattern> .... "
Due to the inclusion of quotes, the exclude pattern is completely
ignored by ctags.

This commit fixes this issue and generates the command line as follows:
ctags -Re -f TAGS --exclude="<pattern>" ...
This does not break the exclude option while still works with spaces in
filenames.