[Fix #672] Don't quit when build Helm source
Instead, only build Helm dwim source when applicable. That is, when the
number of candidates are greater than 1 or no candidate found and the
whole project files are displayed as usual projectile-find-file.

This change also replaces static Helm source with dynamically generated
helm DWIM source. In the future, Helm sources should be changed like
this to adapt to user settings easily. Currently, user has to set fuzzy
matching before loading Helm Projectile, which is inconvenient.