Don't show Projectile modeline when not in project
Currently, Projectile shows any directory it's in. This is not useful
since we don't know whether or not we are in a project directory. This
change makes it only shows project root in modeline when we are actually
in a project, so we know whether we are in a Project or not when
entering a buffer, and also save screen estate.