[Fix #521] Don't add non-projects to project cache
When a user invokes Projectile on a non-project directory, such
directory is cached as 'no-project-root in projectile-project-root-cache,
and the second time when such directory is added with a .projectile file
or anything that Projectile can recognize as a project, it will never
become a project because the cache always suggests Projectile such
directory is not a project. This is a mistake. We should only cache
projects, not non-projects.