Do not pass ignore list to ag in git projects
`ag' supports git repositories and will automatically ignore files
listed in `.gitignore', hence there is no reason to construct the ignore
list.

A benefit of this change is that the command output buffer is much
cleaner. A potential regression that that files the user is ignoring in
projectile, but not ignoring in git (corner case?), will be searched.