Prefer external find-tag implementations
Prior to this commit, projectile attempts to perform completion of tag
searching by building the tags table. This approach is both incredibly
slow but is also subject to truncation of the TAGS file due to
limitations within Emacs.

In this commit, we detect if the user has already loaded a well-known
tags finding function (ggtags, etags-select) and passes off to that
function. We may wish to handle autocompletion ourselves again if the
underlying performance and memory constraint problems can be addressed.

Closes #910 #683 #668 #557 #699