Check prefix *and* suffix when finding test files
When adding `projectile-find-matching-test` support for emacs projects,
I found emacs projects have no standard regarding prefix or suffix---but
`projectile-test-affix` meant a prefix would always hide a suffix if you
listed both.

I believe the fact the result of `projectile-test-affix` was always
tested as both a prefix or a suffix speaks to its lack of value: better
to simply use the result of `projectile-test-prefix` as a prefix or the
result of `projectile-test-suffix` as a suffix directly.

`projectile-find-matching-test` and `projectile-find-matching-file` were
the only users of `projectile-test-affix`, so I simply eliminated it,
after modifying both of them to use the prefixes and suffixes properly.