Ignore cached project roots that point to nonexistent directories
This fixes an awkward bug when the current project is a symlink that
changes. Suppose we have the following filesystem layout:

    /tmp/projectile_project -> /tmp/first_project

If we now delete /tmp/first_project and change the symlink:

    /tmp/projectile_project -> /tmp/second_project

then `projectile-project-root-cache` will contain a key
`"projectile-root-bottom-up-/tmp/projectile-project"` with value
`"/tmp/first_project"`. This will stop `projectile-find-file` from working
as it can't cd to /tmp/first_project.

Instead, we enforce that the string from `projectile-project-root`
points to a directory that currently exists.