Ignore backup files in projectile-get-other-file (#1034)
projectile-get-other-file used to consider backup files of files as the
same extension. So if you called it with "test.c" as current-file and
'("test.h" "test.h~") it would return them both, while the user probably
wouldn't want to edit the backup file. This commit makes sure no backup
files are returned.