Fix file finding in Windows Fossil checkouts
Address #1063.

The Windows fossil.exe outputs \r\n line endings when invoked with the
ls subcommand. In Emacs, this meant that splitting on \0 resulted in a
list of filenames with ^M on the end of each, rendering
projectile-find-file useless in these scenarios. A pipe through dos2unix
before the tr step solves this issue, and where tr is available on a
Windows PATH dos2unix is very likely available as well.