Support setting default compilation dir
Some build systems (eg. meson, autotools and CMake) support building out
out-of-tree. This is already supported in projectile by setting the
`projectile-project-compilation-dir` variable in .dir-locals.el of the
project you prefer an out-of-tree build fr.

However for some build systems (eg. meson) building in-tree is not a
possibility and having to add a .dir-locals.el file to each meson
project is not ergonomic.

Add support for setting a default compilation dir to a project type.