Copy list passed to delete-dups call in projectile-ag
The call to `delete-dups` in `projectile-ag` throws an error if it tries to modify the read-only object `grep-find-ignored-directories`.  This only happens on non-Git projects, since `projectile-ag` only references `grep-find-ignored-directories` if `projectile-project-vcs` does not return `git`.  This PR wraps the list passed to `delete-dups` in a call to `copy-sequence`, which prevents `delete-dups` from modifying `grep-find-ignored-directories`.

I'm not sure if this is the best way to solve the problem.  Please let me know if there is a better solution!

Fixes #1162