Ensure project-roots are directories
Projectile adds projects as directory names, as in "~/project/".
Manually adding a project root that is not a directory name, as
in (projectile-add-known-project "~/project"), works fine, but
Projectile will then add the project root as a directory name, so that
both "~/project" and "~/project/" will be in the list of known
projects.

This change converts "~/project" to "~/project/" before adding it to
the list, thereby avoiding annoying near-duplicates.