Fixes compilation-dir user-error
This fixes the case where projectile-generic-default-command is used by
projectile-compilation-dir() to resolve the compilation-dir. If this is
not set by the project type we throw a user-error, this isn't handled by
projectile-compilation-dir() so we need to handle this special case
gracefully.

The patch adds the special case where the returned default-generic value
returned by the function will purely return nil in that case for
compilation-dir.

A test was added to validate the fix.