[Fix #1232] Switch to a simpler scheme for updating the modeline
Now it's done just once when a file is opened, which should be good
enough for everyone and as fast as possible.

It's funny that after so many years we simply came back to square one.
On the bright side - I can finally delete all the dreaded caching and
optimization code that was added mostly to address this.