[Fix #850] Make it possible to prompt for a project when needed
Now `projectile-require-project-root' gets one more option ('prompt)
that would make certain commands prompt for a project instead of
showing you an error that you're not in a project.

That's just a partial solution for now and it requires
`projectile-ensure-project` to be rolled out to all relevant commands.
As you can see the simple change also required a bit of changes to the
internals, but that's a good thing as some of them were long overdue.