[Fix #1315] Give preference to the project types registered last
We used to keep the project types in a hash table and we didn't have
guarantees about the order in which they its keys would be
returned (or at least I don't think we had), so I've opted to them in
an alist instead.

Now newer registrations take precedence when doing project type look-up.