Improve the naming of the indexing methods
In hindsight I realized that alien and turbo-alien were really bad
names (it's hard to tell them apart and understand what makes them
different) and I've decided to rename the alien method to
hybrid (because this is what it truly is - a hybrid between native and
alien indexing) and now turbo-alien becomes simply alien.

I hope the users will forgive me yet another big behavioural change,
but I think it's fine to do this before having had a stable release with
the new indexing method.