Ignore paths, don’t recursively ignore names of non-existent ignores
find’s “-name” option never matches a path, so binding grep-find-ignored-files
would never allow us to ignore such.  Therefore, .projectile paths like the
following would not ignore anything (except directories unusually bearing
extensions):

-/foo/bar.txt

Currently, it’s only possible to ignore ALL files in a project with a given
name, and only by specifying that name with a leading slash:

-/foo.txt

Specifying a file or directory name without a leading slash will not ignore
anything:

-foo.txt
-foo/

Also, if an ignore was a path to a directory that existed from the root, then
all such paths to directories in a project would be ignored.  However, if the
path was to a direct descendant of the root which was NOT a directory (or did
not even exist), then all FILES in the project with that direct descendant’s
name would be ignored, and if the path was to a 2nd-generation or later
non-directory or non-existent descendant, then nothing would be ignored.

These strange cases stemmed from probable misuse of
grep-find-ignored-directories and grep-find-ignored-files, and greatly limited
what could be ignored.

To ignore paths to files and to get more consistent behavior, it was warranted
to write a special exclusion expression supporting paths and which only ignored
files and directories relative to the root.

Ignores for paths occurring anywhere (not JUST relative to the root) should
probably use the syntax without a leading slash.  Support for that will be
added in a future commit.