Fix cabal project detection
Adjust the wildcard used to detect cabal projects to exclude a
`.cabal` dotfile (or dot-directory).

When using cabal, a `.cabal` directory will exist inside the user's
home directory. Without this change, projectile will interpret that as
a Cabal project existing directly in the home directory, and thus
every file opened that does not belong to a more specific project will
be categorized as a Haskell Cabal project.

Fixes #1391.