Update dotnet project detection
Previously, this function looked only at `.[cf]sproj` files in the _root_ of the
project directory. _However_, this is not how dotnet projects are typically laid
out. Instead, the `.[cf]sproj` file is often located in a directory like
`src/ProjectName/Project.fsproj`. This is because a given repo might contain
_many_ projects, which are then tied together into a root "solution" file.

This commit checks for the solution (`.sln`) file /first/, then does the
existing "same dir" check plus a `src/*/` check for the project files.