Allow multiple root-dirs to be search by projectile-grep
When a project is configured via `.projectile` to include multiple directories
explicitly this results in `projectile-grep` to run over multiple roots.

```
+/foo/bar
+/baz/
```

With a single global *grep* buffer this results in either

a) losing all but the search result of the last root
b) Stalling as the user is asked per root to kill the grep buffer

To avoid this the name of the `*grep*` buffer should to include the current
`root` being search associated with the buffer like `*grep <~/source/foo/bar>*`.