[Fix #1475] Fix unignoring directories
When using hybrid mode and unignoring files that are ignored by git, the
settings in .projectile would not work. The arguments to internal
function projectile-get-repo-ignored-directory were in the wrong
order. So it would return nil instead of the desired list of git ignored
files.

The issue is caught by the projectile-add-unignored test.