mark `projectile--mode-line` as a risky local variable
`projectile--mode-line` is a buffer-local variable that contains the
global and per-buffer mode line content for projectile.  It is rendered
via `(projectile-mode projectile--mode-line)` in `minor-mode-alist`.

When emacs encounters a symbol in rendering the mode-line, it inserts
the string literally.  As documented in [1]:

> Unless SYMBOL is marked as risky (i.e., it has a non-‘nil’
> ‘risky-local-variable’ property), all text properties specified in
> SYMBOL’s value are ignored.

I wanted to use a pretty icon for the fixed part of projectile mode, as
is my habit, but the properties applied to `projectile-mode-line-prefix`
were ignored.

Marking the variable as risky when set with a file-local or
directory-local value allows Emacs to render text properties, including
in my case:

```elisp
(propertize " prj" 'display (create-image ".../icons/projectile.png")
                                          'png nil
                                          :height 18 :width 18
                                          :ascent 'center
                                          :margin 1))
```

This has no harmful effects on any other use of the variables, either
the ones that compose the mode line, or the local variable used to
cache it.

It will, in the unlikely event anyone sets `projectile--mode-line` in
a file/directory local variable, cause Emacs to prompt them to confirm
they are OK with a change that could result in Emacs executing
arbitrary code.  (which, technically, it could do.)

The usual rules for, eg, locally flagging a value safe are still
available, and this is the normal behaviour for any attempt to, eg, set
a hook variable or whatever in file or directory local variables.

[1] https://www.gnu.org/software/emacs/manual/html_node/elisp/Mode-Line-Data.html