projectile-files-via-ext-command: support magic filenames
In 351e0c3b430392c73372f4129b1174c088a934b2 I switched from
`shell-command-to-string` to `call-process-shell-command` so that we
could capture only the stdout, not the stderr.

However, `call-process-shell-command` does not support magic
filenames, which breaks the use of `projectile-files-via-ext-command`
in TRAMP for example.

In this commit we swich to `shell-command`, which supports both
separating stdout and stderr and magic handlers.