Display any errors from projectile files external command
It's useful to see any error messages that these commands return.

This also fixes another issue: previously if pop-up-frames was non-nil and an
error occurred then a new frame would be opened to display the error buffer. But
the error buffer would be immediately deleted by `with-temp-buffer`, leaving an
additional frame displaying the users current buffer.

The use of `save-window-excursion` worked around this issue when only Emacs
windows were in use, which is perhaps why no-one noticed.