Add support for CMake configure, build and test presets.
Introduce a `projectile--cmake-command`, for CMake projects
parametrized on command-type, and route configure, build and test
commands for cmake projects through this.

This command ensures that presets for the given command-type is
supported by the installed version of CMake, that `json-parse-buffer`
is available, that the user has opted in to preset support through
the `projectile-enable-cmake-presets` custom, and that there is at
least one preset available for the given command type.

If all these conditions hold it parses the preset files and prompts
the user to select a preset, or to opt-out of preset configuration,
and creates a manual or preset command based on the users selection.
Else it will revert to manual configuration.